+++ src/v/utils/tests/vint_bench.cc
@@ -306,7 +306,7 @@ static constexpr size_t ITERS = 10000;
 template<typename F>
 ss::future<size_t> co_await_in_loop(F f) {
     perf_tests::start_measuring_time();
-    for (int i = 0; i < ITERS; i++) {
+    for (size_t i = 0; i < ITERS; i++) {
         co_await f();
     }
     perf_tests::stop_measuring_time();
@@ -318,7 +318,7 @@ ss::future<size_t> collect_futures(F f) {
     std::vector<ss::future<>> futs;
     futs.reserve(ITERS);
     perf_tests::start_measuring_time();
-    for (int i = 0; i < ITERS; i++) {
+    for (size_t i = 0; i < ITERS; i++) {
         futs.emplace_back(f());
     }
     perf_tests::stop_measuring_time();
